Adjusts unit tests to conform to new namespace layout.
