+++ ParquetUnitTests/Map/Special Points/ExitPointUnitTest.cs
@@ -3,7 +3,7 @@ using ParquetClassLibrary.Map.SpecialPoints;
 using ParquetClassLibrary.Stubs;
 using Xunit;
 
-namespace ParquetUnitTests.Sandbox
+namespace ParquetUnitTests.Map.SpecialPoints
 {
     public class ExitPointUnitTest
     {
