+++ ParquetUnitTests/Map/Special Points/SpawnPointUnitTest.cs
@@ -1,9 +1,8 @@
-using ParquetClassLibrary.Map;
 using ParquetClassLibrary.Map.SpecialPoints;
 using ParquetClassLibrary.Stubs;
 using Xunit;
 
-namespace ParquetUnitTests.Sandbox
+namespace ParquetUnitTests.Map.SpecialPoints
 {
     public class SpawnPointUnitTest
     {
