+++ ParquetUnitTests/ParquetUnitTests.csproj
@@ -22,14 +22,16 @@
 
   <ItemGroup>
     <Folder Include="Crafting\" />
-    <Folder Include="Sandbox\" />
+    <Folder Include="Map\" />
     <Folder Include="Stubs\" />
-    <Folder Include="Sandbox\Parquets\" />
-    <Folder Include="Sandbox\Special Points\" />
+    <Folder Include="Map\Special Points\" />
     <Folder Include="Utilities\" />
     <Folder Include="Quests\" />
     <Folder Include="Characters\" />
     <Folder Include="Items\" />
+    <Folder Include="Biomes\" />
+    <Folder Include="Parquets\" />
+    <Folder Include="Rooms\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ParquetClassLibrary\ParquetClassLibrary.csproj" />
