+++ ParquetUnitTests/Parquets/FurnishingUnitTest.cs
@@ -3,7 +3,7 @@ using ParquetClassLibrary;
 using ParquetClassLibrary.Parquets;
 using Xunit;
 
-namespace ParquetUnitTests.Sandbox.Parquets
+namespace ParquetUnitTests.Parquets
 {
     public class FurnishingUnitTest
     {
